+++ src/fszmq/Message.fs
@@ -28,6 +28,11 @@ module Message =
   let setOption (message:Message) (messageOption,value) =
     if C.zmq_msg_set(message.Handle,messageOption,value) <> 0 then ZMQ.error()
 
+  /// Sets the given block of option values for the given Message
+  [<Extension;CompiledName("Configure")>]
+  let configure message options =
+    Seq.iter (fun (input:int * int) -> setOption message input) options
+
   /// Returns the content of the given Message
   [<Extension;CompiledName("Data")>]
   let data (message:Message) =
@@ -51,8 +56,8 @@ module Message =
 
   /// For a given message, returns the metadata associated with the given name
   /// as an `Option<string>` where `None` indicates no metadata is present
-  [<Extension;CompiledName("TryGetProperty")>]
-  let tryGetProperty (message:Message) name =
+  [<Extension;CompiledName("TryGetMetadata")>]
+  let tryGetMetadata (message:Message) name =
     match ZMQ.version with
     | Version (m,n,_) 
       when m >= 4 && n >= 1 ->  match C.zmq_msg_gets(message.Handle,name) with
@@ -62,9 +67,12 @@ module Message =
 
   /// For a given message, extracts the metadata associated with the given name,
   /// returning false if no metadata is present (for the given name)
-  [<Extension;CompiledName("TryGetProperty")>]
-  let tryLoadProperty message name ([<Out>]value:byref<string>) =
-    match tryGetProperty message name with
+  ///
+  /// This function is named TryGetMetadata in compiled assemblies.
+  /// If you are accessing the function from a language other than F#, or through reflection, use this name.
+  [<Extension;CompiledName("TryGetMetadata")>]
+  let tryLoadMetadata message name ([<Out>]value:byref<string>) =
+    match tryGetMetadata message name with
     | Some prop ->  value <- prop
                     true
     | None      ->  false